add new RepeatedModel container
