+++ ionics_fits/models/__init__.py
@@ -1,4 +1,4 @@
-from .aggregate_model import AggregateModel
+from .containers import AggregateModel, RepeatedModel
 from .benchmarking import Benchmarking
 from .exponential import Exponential
 from .gaussian import Gaussian
